libmachine/versions: rename to versioncmp
